+++ mission002/YongHoonJJo/js/TodoList.js
@@ -1,5 +1,6 @@
-import { eventKeyboards, counterFilters, dataActions  } from './utils/Contants.js'
+import { counterFilters, dataActions  } from './utils/Contants.js'
 import { createTodoItem, getTodoList, toggleTodoItem, deleteTodoItem } from './utils/apis.js'
+import { renderBySelectedFilterHandler, doByActionHandler, showEditViewHandler, doByEscOrEnterHandler } from './eventHandler/TodoListHandler.js'
 
 function TodoList(selector) {
   const { ALL } = counterFilters
@@ -10,40 +11,9 @@ function TodoList(selector) {
   this.$todoList = document.querySelector(selector)
   this.$filters = document.querySelector('ul.filters')
 
-  this.$filters.addEventListener('click', (e) => {
-    if(e.target === this.$filters) return 
-
-    const aTags = document.querySelectorAll('ul.filters li a')
-    for(const tag of aTags) tag.classList.contains('selected') && tag.classList.remove('selected')
-     
-    this.filter = e.target.className
-    e.target.classList.add('selected')
-
-    this.renderByFilter()
-  })
-
-  this.$todoList.addEventListener('click', async (e) => {
-    const { CHECK, REMOVE } = dataActions
-    const datasetAction = e.target.dataset.action
-    if(!datasetAction) return 
-
-    const [action, id] = datasetAction.split('-')
-    try {
-      switch(action) {
-        case CHECK: await this.toggleState(id); break
-        case REMOVE: await this.removeState(id); break
-      }
-    } catch(e) {
-      console.error({e})
-    }
-  })
-
-  this.$todoList.addEventListener('dblclick', (e) => {
-    const { id } = e.target.dataset
-    if(!id) return 
-
-    this.toggleEditView(id)
-  })
+  this.$filters.addEventListener('click', renderBySelectedFilterHandler.bind(this))
+  this.$todoList.addEventListener('click', doByActionHandler.bind(this))
+  this.$todoList.addEventListener('dblclick', showEditViewHandler.bind(this))
 
   ;(async function() {
     try {
@@ -93,12 +63,12 @@ TodoList.prototype.render = function(items) {
   }, '')
 
   this.$todoList.innerHTML = itemsHtmlString
-  this.renderCounterContainer(items)
+  this.renderCountContainer(items)
 }
 
-TodoList.prototype.renderCounterContainer = function(items) {
-  const $countContainerSpanComponent = document.querySelector('div.count-container > span')
-  $countContainerSpanComponent.innerHTML = `총 <strong>${items.length}</strong> 개`
+TodoList.prototype.renderCountContainer = function(items) {
+  const $countContainerSpan = document.querySelector('div.count-container > span')
+  $countContainerSpan.innerHTML = `총 <strong>${items.length}</strong> 개`
 }
 
 TodoList.prototype.addItem = async function(content) {
@@ -111,28 +81,17 @@ TodoList.prototype.toggleState = async function(id) {
   await this.renderByFilter()
 }
 
-TodoList.prototype.removeState = async function(id) {
+TodoList.prototype.removeItem = async function(id) {
   await deleteTodoItem(this.username, id)
   await this.renderByFilter()
 }
 
 TodoList.prototype.toggleEditView = function(id) {
-  const $liElement = document.querySelector(`li[data-id=${id}]`)
-  const $inputElement = $liElement.querySelector('input.edit')
+  const $li = document.querySelector(`li[data-id="${id}"]`)
+  const $input = $li.querySelector('input.edit')
  
-  $liElement.classList.add('editing')
-
-  $inputElement.addEventListener('keydown', async (e) => {
-    const { ENTER, ESC } = eventKeyboards
-    try {
-      switch(e.key) {
-        case ESC: await this.renderByFilter(); break
-        case ENTER: await this.editContent(id, e.target.value); break
-      }
-    } catch(e) {
-      console.error({e})
-    }
-  })
+  $li.classList.add('editing')
+  $input.addEventListener('keydown', doByEscOrEnterHandler.bind(this, id))
 }
 
 TodoList.prototype.editContent = async function(id, content) {
